型定義を再配置。抽出器だけで１ファイルにする価値がある
