+++ src/features/core/extract/text/eventCommand/types/index.ts
@@ -1,5 +1,6 @@
 export type * from "./bundle";
 export type * from "./extracted";
+export type * from "./extractor";
 export type * from "./pluginCommand";
 export type * from "./result";
 export type * from "./rmmzCommands";
