+++ src/features/core/extract/text/eventCommand/types/pluginCommand.ts
@@ -1,5 +1,8 @@
 import type { PLUGIN_COMMAND_MZ } from "@RpgTypes/libs/eventCommand";
-import type { PluginCommandParameter } from "@RpgTypes/rmmz";
+import type {
+  Command_PluginCommandMZ,
+  PluginCommandParameter,
+} from "@RpgTypes/rmmz";
 
 /**
  * @deprecated Use TextPluginCommandParameter instead
@@ -15,4 +18,8 @@ export interface TextPluginCommandParameter extends PluginCommandParameter<strin
   argName: string;
   argTitle: string;
   speaker?: undefined;
+}
+
+export interface PluginCommandExtractor {
+  extractArgs(command: Command_PluginCommandMZ): TextPluginCommandParameter[];
 }
