+++ src/features/core/extract/text/eventCommand/types/result.ts
@@ -1,29 +1,6 @@
-import type {
-  Command_PluginCommandMZ,
-  Data_CommonEvent,
-  Data_Map,
-  Data_Troop,
-  EventCommand,
-  NoteReadResult,
-} from "@RpgTypes/rmmz";
+import type { NoteReadResult } from "@RpgTypes/rmmz";
 import type { TextPluginCommandParameter } from "./pluginCommand";
 import type { TextCommandParameter } from "./union";
-
-export type GameDataExtractor = EventContainerExtractor & EventCommandExtractor;
-
-export interface EventContainerExtractor {
-  extractMapTexts(rpgMap: Data_Map): ExtractedMapTexts;
-  extractBattleText(troop: Data_Troop): ExtractedBattleEventText[];
-  extractCommonEventText(commons: Data_CommonEvent): ExtractedCommonEventText;
-}
-
-export interface EventCommandExtractor {
-  extractArgs(command: Command_PluginCommandMZ): TextPluginCommandParameter[];
-  extractCommandTexts(
-    commands: ReadonlyArray<EventCommand>,
-  ): TextCommandParameter[];
-  pluginCommandKeys(): string[];
-}
 
 export interface ExtractedBattleEventText<
   T extends TextPluginCommandParameter = TextPluginCommandParameter,
